UCM-14: support arguments for redirecting from both version of httpx in Custom Client
